+++ src/main/java/dev/nicotopia/aoc/graphlib/AStarDataStructure.java
@@ -7,7 +7,7 @@ public interface AStarDataStructure<NodeType> {
      * @param node  The node
      * @param score The new f score
      */
-    void setFScore(NodeType node, int score);
+    void setFScore(NodeType node, long score);
 
     /**
      * Retrieves the last value set by
@@ -17,7 +17,7 @@ public interface AStarDataStructure<NodeType> {
      * @param node The node
      * @return The current f score
      */
-    int getFScore(NodeType node);
+    long getFScore(NodeType node);
 
     /**
      * Sets the cheapest path from a node to a final node that is currently known.
@@ -25,16 +25,16 @@ public interface AStarDataStructure<NodeType> {
      * @param node  The node
      * @param score The new g score
      */
-    void setGScore(NodeType node, int score);
+    void setGScore(NodeType node, long score);
 
     /**
-     * Retrives the last value set by {@link AStarDataStructure#setGScore(Object, int)}
+     * Retrives the last value set by {@link AStarDataStructure#setGScore(Object, long)}
      * or Integer.MAX_VALUE if that never happened.
      * 
      * @param node The node
      * @return The current g score
      */
-    int getGScore(NodeType node);
+    long getGScore(NodeType node);
 
     /**
      * Estimates the distance from a node to a final distance.
@@ -42,7 +42,7 @@ public interface AStarDataStructure<NodeType> {
      * @param node The node
      * @return Estimation of the node's distance to a final node
      */
-    int estimate(NodeType node);
+    long estimate(NodeType node);
 
     /**
      * Checks if a node is final.
